Removed link number overloads. 	Sadly, I don't know what this will break, if anything. On my 	machine the only case is link(p,0), which generates an ambiguity 	error and is easily fixed by explicitly casting the zero to a 	size_t. The benefit is cleaner code and fewer problems for the 	Symbian port, so the change stays until/unless somebody reports 	it as a problem.

+++ ofstream.cc
@@ -92,7 +92,7 @@ ifstream::ifstream (int Fd)
   m_Buffer (255),
   m_File (Fd)
 {
-    link (m_Buffer.data(), 0U);
+    link (m_Buffer.data(), size_t(0));
 }
 
 /// Constructs a stream to read from \p filename.
@@ -102,7 +102,7 @@ ifstream::ifstream (const char* filename, openmode mode)
   m_File (filename, mode)
 {
     clear (m_File.rdstate());
-    link (m_Buffer.data(), 0U);
+    link (m_Buffer.data(), size_t(0));
 }
 
 /// Reads at least \p n more bytes and returns available bytes.
@@ -121,7 +121,7 @@ ifstream::size_type ifstream::underflow (size_type n)
     size_type br = oldPos;
     if (m_Buffer.size() - br < n) {
 	m_Buffer.resize (br + neededFreeSpace);
-	link (m_Buffer.data(), 0U);
+	link (m_Buffer.data(), size_t(0));
     }
     cout.flush();
